==> code02_analysis_of_init/13_HLAfreq.ipynb <==
# Part 13: HLA freq

In this part, we will download and prepare the data from HLAfreq database to compare the HLA distribution in general population and in our cohort. 

pip install --trusted-host pypi.python.org --trusted-host files.pythonhosted.org --trusted-host pypi.org --trusted-host pypi.python.org HLAfreq


import matplotlib.pyplot as plt
import HLAfreq
from HLAfreq import HLAfreq_pymc
base_url = HLAfreq.makeURL(country="Mongolia", standard="g", locus="DQB1")
aftab = HLAfreq.getAFdata(base_url)
caf = HLAfreq.combineAF(aftab)

hdi = HLAfreq_pymc.AFhdi(aftab)
HLAfreq.plotAF(caf, aftab, hdi=hdi, compound_mean=hdi) 

hdi = HLAfreq_pymc.AFhdi(aftab)
HLAfreq.plotAF(caf, aftab, hdi=hdi, compound_mean=hdi) 
base_url = HLAfreq.makeURL(country="Czech Republic", standard="g", locus="DQB1")
aftab = HLAfreq.getAFdata(base_url)
caf = HLAfreq.combineAF(aftab)

Load data from Czech population. 

base_url = HLAfreq.makeURL(country="Czech Republic", standard="g", locus="DQB1")

aftab = HLAfreq.getAFdata(base_url)

import pandas as pd

Save the results. 

country = "Czech Republic"

base_url = HLAfreq.makeURL(country)
aftab = HLAfreq.getAFdata(base_url)
aftab = HLAfreq.decrease_resolution(aftab, 2)
aftab.to_csv("%s_raw.csv" %country, index=False)
